+++ include/vtm/LuaScript.h
@@ -63,7 +63,7 @@ class LuaScript {
   friend VFUDesc *getFUDesc(enum VFUs::FUTypes T);
 
   void initSimpleFU(enum VFUs::FUTypes T, luabind::object FUs,
-                    unsigned *Costs, double *Latencies);
+                    unsigned *Costs, float *Latencies);
 
 public:
